Create cods
In this code, the alarm() function takes the desired alarm time in hours and minutes as parameters. It calculates the time difference between the current time and the alarm time using the std::difftime() function. If the alarm time is in the future, the program sleeps for the calculated time using std::this_thread::sleep_for(). Finally, it prints the alarm message.

In the main() function, the user is prompted to enter the alarm time, and then the alarm() function is called with the provided hours and minutes.

Please note that this code assumes the input values are valid and doesn't handle cases where the user enters incorrect values. You may need to add additional input validation if required.